requirements.txt
python-dotenv==1.0.0
langchain==0.3.10
langchain-experimental
langchain-google-genai
streamlit==1.19.0
InstructorEmbedding==1.0.0
faiss-cpu==1.7.4
langchain-community
google-auth==2.16.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
huggingface-hub==0.13.2
pandas==1.5.3
openai==0.27.0
altair==4.0.0
python-dotenv
